requirements.txt
Flask
UnityPy
Pillow
texture2ddecoder
